More Windows support, better port handling.
On Agent with the usage of the output handler, fix the joining of paths.
On Client get the system temp dir rather than /tmp, and use better
handling with relative/~ directories.

Use a single formatAppFields function to get edit the fields for
creating and updating applications. This fixes an issue where updating
an app to add a custom port would fail because updates wouldn't set the
new sync port.