Implement plugin requirements.
When creating a plugin allow a csv style list, once that gets submitted
the endpoint splits them on "," and stores in the plugin.

When uploading include the requirments struct as json and on the agent
side unmarshal it and verify the os agent is running on is capabale of
running the plugin, if not respond with an error.

A few other bugs found, can't append data in Write methods because
io.Copy and others will check the written length to verify it's correct,
so append before.

Also fix an issue where path expansion was done, just use
formatAppFields and get the LocalPath it responds with.