Running status only after delancey sync added
Commands have optional app ids, if included it'll write to the writers
output, otherwise dev null.

Running status is now set when updating the application is done, that
way it's not possible to run commands and other things requiring an
application on delancey before the sync upload has been called.