Add auto updating to client.
Release zips instead since that's what squirrel requires. Set the
version in the app package.json dynamically when releasing that way the
app package can retrieve the version that's running.

Also update the version of atom-shell we're using, includes a bunch of fixes
related to auto updates. Set the update url to kenmares update check
route, and check it every hour.