Add auto updater for client.
Add --version flag to client that prints the version set at compile
time.

Add install directory to updater program, it lets you modify the
directory to install updates relative to. Also fix issues with
directories in updates. Finally update all paths, not just executables.

The client release script now creates zip uploads to make it easier to
release for Windows users. It also builds the updater and includes it in
the bin dir, finally it creates update tar+gzip downloads for the
updater to use since we're not updating the complete atom app.

The atom main program runs the updater wrapper if it exists and sets the
systems specific installDir(only for Darwin). And fixes an issue with
temp files on Windows.