Fix left over client and infinite errors.
Restarts now use exponential backoff with a limit of 100 retries, the
only thing not counted as a retry is after an update.

Catch signals and forward them to the running process, this fixes an
issue where exiting out of an atom process leaves a client running and
opening it again causes an infinite loop trying to start the
client(which can't because another client is taking the port). Which
resulted in trying ever 5 secs and logs to rollbar causing a huge spike
in network usage.

Atom should kill using SIGINT, SIGKILL doesn't let us clean up and kills
immediately.