Add support for doing batch updates/creates.
This fixes an issue where if you move a large directory into the path being
synced it would cause a delay for changes made during the time the
directories contents are being synced. Instead if there's more than 16
files that have changes then we send them all at once in a .tar.gz to
delancey.